# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RepoPilot",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/github/accounts": {
      "get": {
        "tags": [
          "github"
        ],
        "summary": "List Accounts",
        "operationId": "list_accounts_api_v1_github_accounts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_GithubAccountOut__"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      }
    },
    "/api/v1/github/stars": {
      "get": {
        "tags": [
          "github"
        ],
        "summary": "Get Stars",
        "description": "拉取用户全部 Stars（分页聚合）。\n默认使用 6 小时缓存；?refresh=true 强制更新。",
        "operationId": "get_stars_api_v1_github_stars_get",
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Username"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "强制刷新，忽略缓存",
              "default": false,
              "title": "Refresh"
            },
            "description": "强制刷新，忽略缓存"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_StarsListOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/bindaccount": {
      "post": {
        "tags": [
          "github"
        ],
        "summary": "Bind Account",
        "operationId": "bind_account_api_v1_github_bindaccount_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindGithubBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_GithubAccountOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      }
    },
    "/api/v1/github/accounts/{account_id}": {
      "delete": {
        "tags": [
          "github"
        ],
        "summary": "Unbind Account",
        "operationId": "unbind_account_api_v1_github_accounts__account_id__delete",
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/search": {
      "get": {
        "tags": [
          "github"
        ],
        "summary": "Search Repos",
        "operationId": "search_repos_api_v1_github_search_get",
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_StarRepoOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BindGithubBody": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "pat": {
            "type": "string",
            "maxLength": 256,
            "minLength": 4,
            "title": "Pat"
          }
        },
        "type": "object",
        "required": [
          "username",
          "pat"
        ],
        "title": "BindGithubBody"
      },
...
      "DataResponse_GithubAccountOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GithubAccountOut"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[GithubAccountOut]"
      },
...
      "DataResponse_StarsListOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StarsListOut"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[StarsListOut]"
      },
...
      "DataResponse_dict_": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[dict]"
      },
...
      "DataResponse_list_GithubAccountOut__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/GithubAccountOut"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[list[GithubAccountOut]]"
      },
...
      "DataResponse_list_StarRepoOut__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/StarRepoOut"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[list[StarRepoOut]]"
      },
...
      "GithubAccountOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "bound_at": {
            "type": "string",
            "title": "Bound At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "bound_at"
        ],
        "title": "GithubAccountOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StarRepoOut": {
        "properties": {
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "repo": {
            "type": "string",
            "title": "Repo"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "stars": {
            "type": "integer",
            "title": "Stars",
            "default": 0
          },
          "already_imported": {
            "type": "boolean",
            "title": "Already Imported",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "owner",
          "repo",
          "url"
        ],
        "title": "StarRepoOut"
      },
      "StarsListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StarRepoOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "cache_ttl_hours": {
            "type": "number",
            "title": "Cache Ttl Hours",
            "default": 6.0
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "StarsListOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OptionalAuthBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}